chore: rename data-loader to cli-kintone
