+++ src/__tests__/main.test.ts
@@ -5,10 +5,7 @@ import path from "path";
 const projectRoot = path.resolve(__dirname, "../..");
 const exec = promisify(childProcess.exec);
 const packageJson = require(path.resolve(projectRoot, "package.json"));
-const mainFilePath = path.resolve(
-  projectRoot,
-  packageJson.bin["kintone-data-loader"]
-);
+const mainFilePath = path.resolve(projectRoot, packageJson.bin["cli-kintone"]);
 
 jest.setTimeout(30000);
