+++ src/parsers/parseJson.ts
@@ -1,4 +1,4 @@
-import { RecordForImport } from "../types/data-loader";
+import { RecordForImport } from "../types/cli-kintone";
 
 export const parseJson = (jsonString: string) => {
   return JSON.parse(jsonString) as RecordForImport[];
