+++ src/printers/printAsJson.ts
@@ -1,4 +1,4 @@
-import { RecordForExport } from "../types/data-loader";
+import { RecordForExport } from "../types/cli-kintone";
 
 export const printAsJson: (records: RecordForExport[]) => void = (records) => {
   console.log(JSON.stringify(records, null, 2));
