+++ src/types/cli-kintone.ts
@@ -1,0 +1,5 @@
+export { RecordForExport } from "./cli-kintone/export/record";
+export * as FieldsForExport from "./cli-kintone/export/field";
+
+export { RecordForImport } from "./cli-kintone/import/record";
+export * as FieldsForImport from "./cli-kintone/import/field";
