+++ src/types/cli-kintone/export/record.ts
