+++ src/types/cli-kintone/import/record.ts
