--- src/types/data-loader.ts
@@ -1,5 +1,0 @@
-export { RecordForExport } from "./data-loader/export/record";
-export * as FieldsForExport from "./data-loader/export/field";
-
-export { RecordForImport } from "./data-loader/import/record";
-export * as FieldsForImport from "./data-loader/import/field";
