+++ src/usecase/__tests__/get/fixtures/can_download_files.ts
@@ -1,4 +1,4 @@
-import { RecordForExport } from "../../../../types/data-loader";
+import { RecordForExport } from "../../../../types/cli-kintone";
 import { KintoneRecordForResponse } from "../../../../types/kintone";
 import path from "path";
