+++ src/usecase/__tests__/get/fixtures/can_get_records.ts
@@ -1,4 +1,4 @@
-import { RecordForExport } from "../../../../types/data-loader";
+import { RecordForExport } from "../../../../types/cli-kintone";
 import { KintoneRecordForResponse } from "../../../../types/kintone";
 
 export const input: KintoneRecordForResponse[] = [
